+++ src/components/CardExpirationInput.js
@@ -1,52 +1,13 @@
 import { useState, useRef } from "react";
 import { CARD_EXPIRATION } from "../constants/card";
 import Span from "./Span";
+import useCardExpiration from "../hooks/useCardExpiration";
 
-// TODO : 월(1~12) & 일(1~31)
-export default function CardExpirationInput({
-  className,
-  type,
-  placeholder,
-  maxLength,
-}) {
-  const [inputValue, setInputValue] = useState([null, null]);
-  const dataId = useRef(0); // TODO : nextsibling auto focus & key 값 제어 (useRef)
+export default function CardExpirationInput() {
+  const [inputValue, onChange] = useCardExpiration();
 
-  function isValidExpiration(value) {
-    const pattern_num = /[0-9]/;
-    if (pattern_num.test(value)) return true;
-    if (value === "Backspace" || value === "Tab" || value == "Meta")
-      return true;
-    else return false;
-  }
-
-  function handleKeyDown(event) {
-    const { key } = event;
-    if (!isValidExpiration(key)) {
-      alert("숫자만 입력 가능합니다");
-      event.preventDefault();
-      return false;
-    }
-  }
-
-  function handleChange(event) {
-    const { value } = event.target;
-    console.log(event.target, inputValue, value); // TODO : inputValue 갱신 안되는 버그
-    if (inputValue[0] == event.target) {
-      if (inputValue[0] > 12) {
-        alert("만료 월은 1~12 사이 값만 입력 가능합니다.");
-        event.preventDefault();
-        return false;
-      }
-    } else {
-      if (inputValue[1] > 31) {
-        alert("만료 일은 1~31 사이 값만 입력 가능합니다.");
-        event.preventDefault();
-        return false;
-      }
-    }
-    setInputValue([...inputValue]);
-  }
+  // TODO : nextsibling auto focus & key 값 제어 (useRef)
+  const dataId = useRef(0);
 
   return (
     <div className="input-container">
@@ -58,9 +19,9 @@ export default function CardExpirationInput({
               className="input-basic card-expiration"
               type="text"
               placeholder={index == 0 ? "MM" : "YY"}
-              key={dataId}
-              onKeyDown={handleKeyDown}
-              onChange={handleChange}
+              key={number}
+              onKeyDown={onChange}
+              onChange={onChange}
               defaultValue={number}
               maxLength={CARD_EXPIRATION.MAX_LENGTH}
             ></input>
